# openapi.yaml
openapi: 3.0.3
info:
  title: Facturation.PRO API
  version: "1.0.0"
  description: API REST pour l'outil de facturation Facturation.PRO
  contact:
    email: support@facturation.pro
  license:
    name: Propriétaire
servers:
  - url: https://www.facturation.pro/
paths:
...
  /firms/{firmId}/categories.json:
    parameters:
      - name: firmId
        in: path
        required: true
        schema:
          type: integer
      - name: page
        in: query
        schema:
          type: integer
      - name: title
        in: query
        schema:
          type: string
      - name: status
        in: query
        schema:
          type: integer
    get:
      summary: Liste des catégories
      operationId: listCategories
      responses:
        '200':
          description: Liste des catégories
          headers:
            X-Pagination:
              description: Informations de pagination
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Category'
    post:
      summary: Créer une catégorie
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - status
              properties:
                title:
                  type: string
                status:
                  type: integer
      responses:
        '201':
          description: Catégorie créée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
  /firms/{firmId}/categories/{id}.json:
    parameters:
      - name: firmId
        in: path
        required: true
        schema:
          type: integer
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Détails d'une catégorie
      operationId: getCategory
      responses:
        '200':
          description: Détails de la catégorie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
    patch:
      summary: Modifier une catégorie
      operationId: updateCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                status:
                  type: integer
      responses:
        '200':
          description: Catégorie modifiée
    delete:
      summary: Supprimer une catégorie
      operationId: deleteCategory
      responses:
        '200':
          description: Catégorie supprimée

  # Clients
  /firms/{firmId}/customers.json:
    parameters:
      - name: firmId
        in: path
        required: true
        schema:
          type: integer
      - name: page
        in: query
        schema:
          type: integer
      - name: mode
        in: query
        schema:
          type: string
          enum: [all, company, individual, archived]
      - name: api_id
        in: query
        schema:
          type: integer
      - name: api_custom
        in: query
        schema:
          type: string
      - name: company
        in: query
        schema:
          type: string
      - name: last_name
        in: query
        schema:
          type: string
      - name: email
        in: query
        schema:
          type: string
      - name: category_id
        in: query
        schema:
          type: integer
      - name: with_sepa
        in: query
        schema:
          type: integer
      - name: account_code
        in: query
        schema:
          type: string
      - name: accounting_entry
        in: query
        schema:
          type: string
      - name: sort
        in: query
        schema:
          type: string
          enum: [asc, desc]
      - name: order
        in: query
        schema:
          type: string
          enum: [last_invoice, last_paid, created, updated]
    get:
      summary: Liste des clients
      operationId: listCustomers
      responses:
        '200':
          description: Liste des clients
          headers:
            X-Pagination:
              description: Informations de pagination
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
    post:
      summary: Créer un client
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                company_name:
                  type: string
                individual:
                  type: boolean
                first_name:
                  type: string
                last_name:
                  type: string
                street:
                  type: string
                city:
                  type: string
                zip_code:
                  type: string
                country:
                  type: string
                phone:
                  type: string
                fax:
                  type: string
                email:
                  type: string
                mobile:
                  type: string
                website:
                  type: string
                notes:
                  type: string
                vat_number:
                  type: string
                siret:
                  type: string
                category_id:
                  type: integer
                account_code:
                  type: string
                accounting_entry:
                  type: string
                currency:
                  type: string
                language:
                  type: string
                pay_before:
                  type: string
                reference:
                  type: string
                field1:
                  type: string
                field2:
                  type: string
                field3:
                  type: string
                field4:
                  type: string
                field5:
                  type: string
      responses:
        '201':
          description: Client créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
  /firms/{firmId}/customers/{id}.json:
    parameters:
      - name: firmId
        in: path
        required: true
        schema:
          type: integer
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: with_sepa
        in: query
        schema:
          type: integer
    get:
      summary: Détails d'un client
      operationId: getCustomer
      responses:
        '200':
          description: Détails du client
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
    patch:
      summary: Modifier un client
      operationId: updateCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                company_name:
                  type: string
                individual:
                  type: boolean
                first_name:
                  type: string
                last_name:
                  type: string
                street:
                  type: string
                city:
                  type: string
                zip_code:
                  type: string
                country:
                  type: string
                phone:
                  type: string
                fax:
                  type: string
                email:
                  type: string
                mobile:
                  type: string
                website:
                  type: string
                notes:
                  type: string
                vat_number:
                  type: string
                siret:
                  type: string
                category_id:
                  type: integer
                account_code:
                  type: string
                accounting_entry:
                  type: string
                currency:
                  type: string
                language:
                  type: string
                pay_before:
                  type: string
                reference:
                  type: string
                field1:
                  type: string
                field2:
                  type: string
                field3:
                  type: string
                field4:
                  type: string
                field5:
                  type: string
      responses:
        '200':
          description: Client modifié
    delete:
      summary: Supprimer un client
      operationId: deleteCustomer
      responses:
        '200':
          description: Client supprimé
  /firms/{firmId}/customers/{id}/archive.json:
    parameters:
      - name: firmId
        in: path
        required: true
        schema:
          type: integer
      - name: id
        in: path
        required: true
        schema:
          type: integer
    post:
      summary: Archiver un client
      operationId: archiveCustomer
      responses:
        '200':
          description: Client archivé
  /firms/{firmId}/customers/{id}/unarchive.json:
    parameters:
      - name: firmId
        in: path
        required: true
        schema:
          type: integer
      - name: id
        in: path
        required: true
        schema:
          type: integer
    post:
      summary: Restaurer un client
      operationId: unarchiveCustomer
      responses:
        '200':
          description: Client restauré
  /firms/{firmId}/customers/{id}/upload.json:
    parameters:
      - name: firmId
        in: path
        required: true
        schema:
          type: integer
      - name: id
        in: path
        required: true
        schema:
          type: integer
      - name: filename
        in: query
        schema:
          type: string
    post:
      summary: Ajouter un fichier à un client
      operationId: uploadCustomerFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                upload_file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Fichier ajouté
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileAttachment'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
          description: Identifiant
        title:
          type: string
          description: Libellé
        status:
          type: integer
          description: Affectation
        created_at:
          type: string
          format: date-time
          description: Créé le
        updated_at:
          type: string
          format: date-time
          description: Modifié le
...
    FileAttachment:
      type: object
      properties:
        id:
          type: integer
        purchase_id:
          type: integer
        quote_id:
          type: integer
        invoice_id:
          type: integer
        customer_id:
          type: integer
        document_name:
          type: string
        document_size:
          type: integer
        download_url:
          type: string
        title:
          type: string
        visible:
          type: boolean